Day 4 Arrays & Loops
adding functions :
sum(numbers) : a function that takes an array of numbers and returns their sum.

countEven(numbers) : a function that takes an array of numbers and returns the count of even numbers.

double(numbers) : a function that takes an array of numbers and returns a new array where each element is doubled

sockMerchant(socks)Given an array of integers representing the color of each sock, determine how many pairs of socks with matching colors there are.